Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/html/infrastructure/urls/resolving-urls/query-encoding/windows-1252-expected.txt

CONSOLE MESSAGE: WebSocket connection to 'ws://localhost:49001/echo-query?%C3%A5' failed: WebSocket is closed before the connection is established.
CONSOLE MESSAGE: Did not parse stylesheet at 'http://localhost:8800/html/infrastructure/urls/resolving-urls/query-encoding/resources/stash.py?q=%C3%A5&action=put&id=d6ee874b-a162-41cc-b22f-5ab9030f5cf9' because non CSS MIME types are not allowed in strict mode.
CONSOLE MESSAGE: Did not parse stylesheet at 'http://localhost:8800/html/infrastructure/urls/resolving-urls/query-encoding/resources/stash.py?q=%C3%A5&action=put&id=96e0db0a-ac82-483f-b68d-c8074d6fce5b' because non CSS MIME types are not allowed in strict mode.
CONSOLE MESSAGE: Did not parse stylesheet at 'http://localhost:8800/html/infrastructure/urls/resolving-urls/query-encoding/resources/stash.py?q=%C3%A5&action=put&id=f6e87bbb-a25a-4dd1-9321-22fc0ad31a2a' because non CSS MIME types are not allowed in strict mode.
CONSOLE MESSAGE: ApplicationCache is deprecated. Please use ServiceWorkers instead.
CONSOLE MESSAGE: ApplicationCache is deprecated. Please use ServiceWorkers instead.
CONSOLE MESSAGE: ApplicationCache is deprecated. Please use ServiceWorkers instead.

Harness Error (TIMEOUT), message = null

PASS load nested browsing context <frame src>
PASS load nested browsing context <iframe src>
TIMEOUT load nested browsing context <object data> Test timed out
TIMEOUT load nested browsing context <embed src> Test timed out
PASS loading css <link>
PASS loading js <script>
PASS loading image <img src>
TIMEOUT loading image <embed src> Test timed out
TIMEOUT loading image <object data> Test timed out
PASS loading image <input src>
PASS loading image <video poster>
TIMEOUT loading video <video> Test timed out
TIMEOUT loading video <video><source> Test timed out
TIMEOUT loading video <audio> Test timed out
TIMEOUT loading video <audio><source> Test timed out
PASS loading webvtt <track>
PASS submit form <form action>
PASS submit form <input formaction>
PASS submit form <button formaction>
FAIL <base href> assert_true: expected substring %E5 got http://localhost:8800/html/infrastructure/urls/resolving-urls/query-encoding/resources/resource.py?q=%C3%A5&encoding=windows-1252&type= expected true got false
PASS Worker constructor
FAIL SharedWorker constructor Can't find variable: SharedWorker
PASS EventSource constructor
PASS EventSource#url
FAIL XMLDocument#load() doc.load is not a function. (In 'doc.load(input_url_svg)', 'doc.load' is undefined)
PASS window.open()
FAIL <a>.search assert_true: href content attribute expected substring %E5 got http://localhost:8800/html/infrastructure/urls/resolving-urls/query-encoding/resources/resource.py?q=%C3%A5&encoding=windows-1252&type=html expected true got false
FAIL <area>.search assert_true: href content attribute expected substring %E5 got http://localhost:8800/html/infrastructure/urls/resolving-urls/query-encoding/resources/resource.py?q=%C3%A5&encoding=windows-1252&type=html expected true got false
FAIL history.pushState assert_equals: url was resolved against the iframe's URL instead of the settings object's API base URL expected -1 but got 76
FAIL history.replaceState assert_equals: url was resolved against the iframe's URL instead of the settings object's API base URL expected -1 but got 76
PASS SVG <a>
TIMEOUT SVG <feImage>
FAIL SVG <image> assert_equals: expected "%E5" but got "%C3%A5"
PASS SVG <use>
FAIL XMLHttpRequest#open() assert_equals: expected "%C3%A5" but got "%E5"
PASS importScripts() in a dedicated worker
FAIL Worker() in a dedicated worker assert_equals: expected "%C3%A5" but got "ReferenceError: Can't find variable: Worker"
FAIL SharedWorker() in a dedicated worker assert_equals: expected "%C3%A5" but got "ReferenceError: Can't find variable: SharedWorker"
FAIL importScripts() in a shared worker Can't find variable: SharedWorker
FAIL Worker() in a shared worker Can't find variable: SharedWorker
FAIL SharedWorker() in a shared worker Can't find variable: SharedWorker
PASS WebSocket constructor
PASS WebSocket#url
PASS Parsing cache manifest (CACHE)
PASS Parsing cache manifest (FALLBACK)
PASS Parsing cache manifest (NETWORK)
FAIL CSS <link> (windows-1252) #<id> { background-image:<url> } assert_equals: expected "%C3%A5" but got "%E5"
PASS CSS <link> (utf-8) #<id> { background-image:<url> }
FAIL CSS <style> #<id> { background-image:<url> } assert_equals: expected "%C3%A5" but got "%E5"
FAIL CSS <link> (windows-1252) #<id> { border-image-source:<url> } assert_equals: expected "%C3%A5" but got "%E5"
PASS CSS <link> (utf-8) #<id> { border-image-source:<url> }
FAIL CSS <style> #<id> { border-image-source:<url> } assert_equals: expected "%C3%A5" but got "%E5"
FAIL CSS <link> (windows-1252) #<id>::before { content:<url> } assert_equals: expected "%C3%A5" but got "%E5"
PASS CSS <link> (utf-8) #<id>::before { content:<url> }
FAIL CSS <style> #<id>::before { content:<url> } assert_equals: expected "%C3%A5" but got "%E5"
FAIL CSS <link> (windows-1252) @font-face { font-family:<id>; src:<url> } #<id> { font-family:<id> } assert_equals: expected "%C3%A5" but got "%E5"
PASS CSS <link> (utf-8) @font-face { font-family:<id>; src:<url> } #<id> { font-family:<id> }
FAIL CSS <style> @font-face { font-family:<id>; src:<url> } #<id> { font-family:<id> } assert_equals: expected "%C3%A5" but got "%E5"
FAIL CSS <link> (windows-1252) #<id> { display:list-item; list-style-image:<url> } assert_equals: expected "%C3%A5" but got "%E5"
PASS CSS <link> (utf-8) #<id> { display:list-item; list-style-image:<url> }
FAIL CSS <style> #<id> { display:list-item; list-style-image:<url> } assert_equals: expected "%C3%A5" but got "%E5"
PASS CSS <link> (windows-1252) @import <url>;
PASS CSS <link> (utf-8) @import <url>;
PASS CSS <style> @import <url>;
FAIL CSS <link> (windows-1252) #<id> { cursor:<url>, pointer } assert_equals: expected "%C3%A5" but got "%E5"
PASS CSS <link> (utf-8) #<id> { cursor:<url>, pointer }
FAIL CSS <style> #<id> { cursor:<url>, pointer } assert_equals: expected "%C3%A5" but got "%E5"
FAIL <?xml-stylesheet?> (CSS) assert_equals: expected "\"%C3%A5\"" but got "\"%E5\""
PASS URL constructor, url
PASS URL constructor, base
PASS Scheme ftp (getting <a>.href)
PASS Scheme file (getting <a>.href)
PASS Scheme gopher (getting <a>.href)
PASS Scheme http (getting <a>.href)
PASS Scheme https (getting <a>.href)
PASS Scheme ws (getting <a>.href)
PASS Scheme wss (getting <a>.href)
PASS Scheme mailto (getting <a>.href)
PASS Scheme data (getting <a>.href)
PASS Scheme javascript (getting <a>.href)
PASS Scheme ftps (getting <a>.href)
PASS Scheme httpbogus (getting <a>.href)
PASS Scheme bitcoin (getting <a>.href)
PASS Scheme geo (getting <a>.href)
PASS Scheme im (getting <a>.href)
PASS Scheme irc (getting <a>.href)
PASS Scheme ircs (getting <a>.href)
PASS Scheme magnet (getting <a>.href)
PASS Scheme mms (getting <a>.href)
PASS Scheme news (getting <a>.href)
PASS Scheme nntp (getting <a>.href)
PASS Scheme sip (getting <a>.href)
PASS Scheme sms (getting <a>.href)
PASS Scheme smsto (getting <a>.href)
PASS Scheme ssh (getting <a>.href)
PASS Scheme tel (getting <a>.href)
PASS Scheme urn (getting <a>.href)
PASS Scheme webcal (getting <a>.href)
PASS Scheme wtai (getting <a>.href)
PASS Scheme xmpp (getting <a>.href)
PASS Scheme web+http (getting <a>.href)
